change: using User id from Token  - requests get user id from the http only token,  not from the request params

+++ backend/controllers/controllers_todos.ts
@@ -18,7 +18,7 @@ export const getTodosByUserId = async (req: Request, res: Response) => {
       if (error) return res.sendStatus(403)
 
       let sql:string = `SELECT * FROM ${TodosTable} WHERE user_id = ?`;
-      const todos: Todo[] = await useMySql(sql, [req.params.id]);
+      const todos: Todo[] = await useMySql(sql, [authData.userId]);
       res.status(200).json(todos);
 
     })
@@ -30,7 +30,7 @@ export const newTodo = async (req: Request, res: Response) => {
     jwt.verify((req as TokenRequest).token, process.env.secret_key!, async (error: any, authData: any) => {
       if (error) return res.sendStatus(403)
 
-      let newTodo: Todo = {user_id: req.body.user_id, text: req.body.text, isDone: 0};
+      let newTodo: Todo = {user_id: authData.userId, text: req.body.text, isDone: 0};
       let sql:string = `INSERT INTO ${TodosTable} (user_id, text, isDone) VALUES (?,?,?)`;
       await useMySql(sql, [newTodo.user_id, newTodo.text, newTodo.isDone]);
       res.status(201).json({ message: "Todo added" });
@@ -45,18 +45,18 @@ export const updateTodo = async (req: Request, res: Response) => {
       if (error) return res.sendStatus(403)
 
       if (req.body.text) { // update text
-        let sqlText:string = `UPDATE ${TodosTable} SET text = ? WHERE id = ?`;
-        await useMySql(sqlText, [req.body.text, req.params.id]);
+        let sqlText:string = `UPDATE ${TodosTable} SET text = ? WHERE id = ? AND user_id = ?`;
+        await useMySql(sqlText, [req.body.text, req.params.id,authData.userId]);
         res.status(201).json({ message: "Todo updated" });
 
       } else {  // update isDone
-        let sqlGET:string = `SELECT * FROM ${TodosTable} WHERE id = ?`;
-        const todos: Todo[] = await useMySql(sqlGET, [req.params.id]);
+        let sqlGET:string = `SELECT * FROM ${TodosTable} WHERE id = ? AND user_id = ?`;
+        const todos: Todo[] = await useMySql(sqlGET, [req.params.id,authData.userId]);
 
         let update:number = todos[0].isDone === 1 ? 0 : 1;
 
-        let sql:string = `UPDATE ${TodosTable} SET isDone = ? WHERE id = ?`;
-        await useMySql(sql, [update, req.params.id]);
+        let sql:string = `UPDATE ${TodosTable} SET isDone = ? WHERE id = ? AND user_id = ?`;
+        await useMySql(sql, [update, req.params.id, authData.userId]);
         res.status(200).json({ message: "Todo updated" });
       }
 
@@ -70,8 +70,8 @@ export const deleteTodoByTodoId = async (req: Request, res: Response) => {
     jwt.verify((req as TokenRequest).token, process.env.secret_key!, async (error: any, authData: any) => {
       if (error) return res.sendStatus(403)
 
-      let sql:string = `DELETE FROM ${TodosTable} WHERE id = ?`;
-      await useMySql(sql, [req.params.id]);
+      let sql:string = `DELETE FROM ${TodosTable} WHERE id = ? AND user_id = ?`;
+      await useMySql(sql, [req.params.id,authData.userId]);
       res.status(200).json({ message: "Todo deleted" });
 
     })
